+++ NEWS.md
@@ -1,5 +1,8 @@
 # MiscMetabar 0.11.1 (in development) 
 
+- Add param `suffix` to `add_blast_info()` allowing multiple use of the function on the same phyloseq object (e.g. in order to used different database)
+- Add param `return_DNAStringSet` to `write_temp_fasta()` function to return a DNAStringSet object in place of a temporary file. 
+- Add a vignette pkgnet-report.
 
 # MiscMetabar 0.11